+++ src/disclosures/js/views/financial-view.js
@@ -154,7 +154,8 @@ var financialView = {
     this.updateLeftovers( values, $leftovers );
     this.updatePrivateLoans( values, $privateLoans );
     this.updateRemainingCostContent();
-    metricView.updateDebtBurdenDisplay( values );
+    metricView.updateDebtBurden( values );
+    metricView.updateMonthlyPayment();
   },
 
   /**